test: split agent tools suite into document and workbook files
- rename test_office_assistant_agent_tools.py to test_office_assistant_document_tools.py and drop workbook-only cases and imports so Word coverage stays focused

- add test_office_assistant_workbook_tools.py to collect workbook agent tool cases, MCP write_rows cases, and workbook loader cases in one place

- keep the cross-domain MCP registration cases in the document file since workbook tools are registered as a supplement to document tools
